+++ GameEngine/Tools/AssetsEditor/Systems/AssetsAndComponentRelationsSystem.hpp
@@ -15,9 +15,9 @@ namespace AGE
 			EntityFilter _rigidBodies;
 			EntityFilter _meshRenderer;
 
-			virtual void updateBegin(double time);
-			virtual void updateEnd(double time);
-			virtual void mainUpdate(double time);
+			virtual void updateBegin(float time);
+			virtual void updateEnd(float time);
+			virtual void mainUpdate(float time);
 			virtual bool initialize();
 		};
 	}
